requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
geographiclib==1.50
geopy==2.0.0
Pillow==7.1.2
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.31
whitenoise==5.2.0
